:truck: move ComplementaryCertification to configuration
