+++ api/src/certification/configuration/domain/models/ComplementaryCertification.js
@@ -1,6 +1,6 @@
 import { ComplementaryCertificationKeys } from '../../../shared/domain/models/ComplementaryCertificationKeys.js';
 
-class ComplementaryCertification {
+export class ComplementaryCertification {
   constructor({ id, label, key }) {
     this.id = id;
     this.label = label;
@@ -8,5 +8,3 @@ class ComplementaryCertification {
     this.hasComplementaryReferential = key !== ComplementaryCertificationKeys.CLEA;
   }
 }
-
-export { ComplementaryCertification };
