+++ api/tests/tooling/domain-builder/factory/certification/complementary-certification/build-complementary-certification.js
@@ -1,4 +1,4 @@
-import { ComplementaryCertification } from '../../../../../../src/certification/complementary-certification/domain/models/ComplementaryCertification.js';
+import { ComplementaryCertification } from '../../../../../../src/certification/configuration/domain/models/ComplementaryCertification.js';
 import { ComplementaryCertificationKeys } from '../../../../../../src/certification/shared/domain/models/ComplementaryCertificationKeys.js';
 
 const buildComplementaryCertification = function ({
